Implemented setup validity check.

+++ player.cc
@@ -1,5 +1,5 @@
 #include "player.h"
 using namespace std;
 
-Player::Player(Colour c, int score): c{c}, score{score} {}
+Player::Player(Colour c, double score): c{c}, score{score} {}
 Player::~Player() {}
